Fedora-centric file added
Fedora likes to not wait for dhclient to return before it tries to
compile a list of files to scan. So we force it to wait longer than
normal (2 -> 5 seconds) in the hope the files are created successfully
to allow a successful password change.
This only happens when leases are _not_ already in existence and do
_not_ have a DHCP host listed... so it should not impact performance in
the long run but definitely improves reliability.